feature/#5: 좋아요 두 번 생성 방지 및 유효성 검사 추가

+++ app/src/models/love/loveService.js
@@ -10,22 +10,55 @@ class LoveService {
   async getLove() {
     const userNo = this.body.user_no;
     const boardNo = this.params.board_no;
-    try {
-      const response = await LoveRepository.getLove(userNo, boardNo);
-      return response;
-    } catch (err) {
-      return { msg: err };
+    let error = await LoveRepository.isBoardNo(boardNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 게시글이 존재하지 않습니다.",
+        statuscode: 404,
+      };
+    }
+    error = await LoveRepository.isUserNo(userNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 유저가 존재하지 않습니다.",
+        statuscode: 404,
+      };
+    }
+    const duplication = await LoveRepository.duplication(boardNo, userNo);
+    if (!duplication[0][0]) {
+      await LoveRepository.getLove(userNo, boardNo);
+      return {statuscode: 201};
+    } else {
+      return {
+        error: "Bad Request",
+        message: "좋아요는 한 번만 가능합니다.",
+        statuscode: 400,
+      };
     }
   }
   async deleteLove() {
     const userNo = this.body.user_no;
     const boardNo = this.params.board_no;
-    try {
-      const response = await LoveRepository.deleteLove(userNo, boardNo);
-      return response;
-    } catch (err) {
-      return { msg: err };
+    let error = await LoveRepository.isBoardNo(boardNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 게시글이 존재하지 않습니다.",
+        statuscode: 404,
+      };
     }
+    error = await LoveRepository.isUserNo(userNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 유저가 존재하지 않습니다.",
+        statuscode: 404,
+      };
+    }
+    await LoveRepository.deleteLove(userNo, boardNo);
+    return {statuscode: 200};
   }
 }
